day-83/requirements.txt
Flask==3.0.3
flask_mail==0.10.0
python-dotenv==1.0.1
